+++ dcp-common/src/main/java/com/vmware/dcp/common/ServiceHost.java
@@ -2900,14 +2900,7 @@ public class ServiceHost {
     }
 
     public void sendRequest(Operation op) {
-        if (op.isJoined()) {
-            for (Operation currentOp : op.getJoinedOperations()) {
-                prepareRequest(currentOp);
-            }
-        } else {
-            prepareRequest(op);
-        }
-
+        prepareRequest(op);
         traceOperation(op);
 
         if (this.isStopping()) {